# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/planos": {
            "get": {
                "tags": [
                    "Planos"
                ],
                "summary": "Buscar planos",
                "description": "Todos os planos",
                "operationId": "5df044e8ed170e3741410bb3e66c5cf8",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "The response data",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "nome": {
                                                        "type": "string"
                                                    },
                                                    "preco": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "descricao": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "errcode": 1,
                                        "errmsg": "ok",
                                        "data": []
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Planos"
                ],
                "summary": "Novo Plano",
                "operationId": "9033a6ba719bd3e482b957d254a61a35",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "header",
                        "description": "nome do plano",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descricao",
                        "in": "header",
                        "description": "descricao do plano",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "preco",
                        "in": "header",
                        "description": "preco do plano",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Plano registrado com sucesso"
                    },
                    "422": {
                        "description": "Erro na validação"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/planos/{id}": {
            "delete": {
                "tags": [
                    "Planos"
                ],
                "summary": "Excluir um plano",
                "operationId": "b5182f9c814f6704dd4847a375ba76d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do plano",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Plano exluido com sucesso"
                    },
                    "422": {
                        "description": "Erro na validação"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Planos"
                ],
                "summary": "Atualizar um plano",
                "operationId": "61aa99b59c11199c694c171c5d0e6efd",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "descricao",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "preco",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do plano",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Assinatura atualizada com sucesso"
                    },
                    "422": {
                        "description": "Erro na validação"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}